requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
face-recognition==1.3.0
